# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Task Manager API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter JWT token directly (without 'Bearer ' prefix). Example: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...'"
        }
    },
    "security": [],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login and get JWT token",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "summary": "List tasks",
                "parameters": [
                    {
                        "name": "completed",
                        "in": "query",
                        "type": "boolean"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks"
                    }
                }
            },
            "post": {
                "summary": "Create a task",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task created"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "summary": "Get a task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "summary": "Update a task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "completed": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    }
}